badge周りの処理を追加
